Add the functionality for closing the game, fix the click counter
The game can be closed with the close button at any time, then reopened to the intro screen and played again. The click counter was fixed to account for the fact that a double click is effectively counted as three clicks: two regular clicks plus one double click (which gives the user 3 points instead of 2 even though they clicked twice). Also: reverted to using the hide class for the results pane.